+++ back/src/main/java/com/realive/RealiveApplication.java
@@ -10,5 +10,4 @@ public class RealiveApplication {
 		SpringApplication.run(RealiveApplication.class, args);
 	}
 
-
 }
